subsidy/tests/requirements-test.txt
pytest==7.4.0
pytest-cov==4.1.0
httpx==0.24.1
python-jose==3.3.0
python-multipart==0.0.6
fastapi==0.103.1
pydantic==2.3.0
pydantic-settings==2.0.3
uvicorn==0.23.2
